Impedir que se pueda eliminar al administrador y a los usuarios con reservas pendientes

+++ FitCenterBooking/src/main/java/com/salesianostriana/dam/fitcenterbooking/controller/ControllerUsuario.java
@@ -1,5 +1,8 @@
 package com.salesianostriana.dam.fitcenterbooking.controller;
 
+import java.time.LocalDateTime;
+import java.util.List;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,9 +10,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.salesianostriana.dam.fitcenterbooking.model.Reserva;
 import com.salesianostriana.dam.fitcenterbooking.model.Usuario;
 import com.salesianostriana.dam.fitcenterbooking.security.RolesUsuario;
+import com.salesianostriana.dam.fitcenterbooking.service.ServiceReserva;
 import com.salesianostriana.dam.fitcenterbooking.service.ServiceUsuario;
 
 import lombok.RequiredArgsConstructor;
@@ -19,6 +25,8 @@ import lombok.RequiredArgsConstructor;
 public class ControllerUsuario {
 
 	private final ServiceUsuario service;
+	private final ServiceReserva serviceReserva;
+
     private final PasswordEncoder encoder;
 
 
@@ -30,13 +38,13 @@ public class ControllerUsuario {
 	}
 	
 	@GetMapping("/login")
-	public String login (Model model) {
+	public String verLogin (Model model) {
 		
 		return "login";
 	}
 	
 	@GetMapping("/registro")
-	public String registro (Model model) {
+	public String verRegistro (Model model) {
 		
 		Usuario nuevoUsuario = new Usuario();
         
@@ -55,15 +63,36 @@ public class ControllerUsuario {
     }
 	
 	@GetMapping("/usuarios/eliminar/{id}")
-	public String eliminarActividad(@PathVariable("id") Long id) {
+	public String eliminarUsuario(@PathVariable("id") Long id, RedirectAttributes redirect) {
 		
-		service.deleteById(id); 
+		Usuario usuarioABorrar = service.buscarPorID(id);
+		
+		if (usuarioABorrar.getEmail().equals("admin@admin.com")) {
+	        redirect.addFlashAttribute("error", "No se puede eliminar al administrador del sistema por defecto");
+	        return "redirect:/usuarios";
+	    }
+		
+		List<Reserva> susReservas = serviceReserva.listarReservasUsuario(id);
+	    
+	    boolean tieneReservasActivas = false;
+	            
+	    for(Reserva r : susReservas) {
+	    	if(r.getFecha().isAfter(LocalDateTime.now())) {
+	    		tieneReservasActivas = true;
+	    	}
+	    }
+	    if (tieneReservasActivas) {
+	        redirect.addFlashAttribute("errorUsuario", "No se puede eliminar al usuario porque tiene reservas pendientes.");
+	        return "redirect:/usuarios";
+	    }
+	    
+	    service.deleteById(id);
 		
 		return "redirect:/usuarios";
 	}
 	
 	@GetMapping("/usuarios/editar/{id}")
-	public String editarActividad(@PathVariable("id") Long id, Model model) {
+	public String editarUsuario(@PathVariable("id") Long id, Model model) {
 		
 		Usuario u = service.buscarPorID(id);
         model.addAttribute("usuario", u);
